+++ src/serviceWorker.ts
@@ -1,0 +1,8 @@
+/// <reference types="@sveltejs/kit" />
+/// <reference no-default-lib="true"/>
+/// <reference lib="esnext" />
+/// <reference lib="webworker" />
+
+// declare const self: ServiceWorkerGlobalScope;
+
+export {};